+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Link from widget to corresponding Matomo report if available (#7)
 - Widget "Browser Plugins" (#12)
 - Widget for Custom Dimensions (#6)
+- Widget for Content Names (#8)
 
 ### Changed
 - Use selectCheckBox in site configuration for active widget selection instead of checkboxLabeledToggle, an upgrade wizard is available (#10)